+++ libs/ionic-react/src/utils/versions.ts
@@ -5,5 +5,5 @@ export const ionicReactRouterVersion = '5.0.7';
 export const ioniconsVersion = '5.0.1';
 
 export const testingLibraryJestDomVersion = '5.5.0';
-export const testingLibraryUserEventVersion = '8.1.3';
+export const testingLibraryUserEventVersion = '10.0.1';
 export const testingLibraryCypressVersion = '6.0.0';