+++ libs/ionic-react/CHANGELOG.md
@@ -8,6 +8,7 @@
 - import `@testing-library/jest-dom` commands for unit tests
 - upgrade `@testing-library/jest-dom` to 5.5.0
 - upgrade `@testing-library/cypress` to 6.0.0
+- upgrade `@testing-library/user-event` to 10.0.1
 
 # 1.0.2